--- a/Project.toml
+++ b/Project.toml
@@ -4,6 +4,7 @@ authors = ["Takafumi Arakaki <aka.tkf@gmail.com>", "Jan Weidner <jw3126@gmail.co
 version = "0.1.37"
 
 [deps]
+ArgCheck = "dce04be8-c92d-5529-be00-80e4d2c0e197"
 CompositionsBase = "a33af91c-f02d-484b-be07-31d278c5ca2b"
 ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -12,7 +13,6 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [weakdeps]
 AxisKeys = "94b1ba4f-4ee9-5380-92f1-94cde586c3c5"
@@ -30,6 +30,7 @@ AccessorsStructArraysExt = "StructArrays"
 AccessorsUnitfulExt = "Unitful"
 
 [compat]
+ArgCheck = "2"
 AxisKeys = "0.2"
 CompositionsBase = "0.1"
 ConstructionBase = "1.5"
@@ -42,7 +43,6 @@ Markdown = "1"
 Requires = "0.5, 1.0"
 StaticArrays = "1"
 StructArrays = "0.6"
-Test = "1"
 Unitful = "1"
 julia = "1.6"