--- a/src/Accessors.jl
+++ b/src/Accessors.jl
@@ -19,7 +19,6 @@ include("testing.jl")
 # always included for now
 include("../ext/AccessorsDatesExt.jl")
 include("../ext/AccessorsLinearAlgebraExt.jl")
-include("../ext/AccessorsTestExt.jl")
 
 function __init__()
     @static if !isdefined(Base, :get_extension)