This PR replaces the use of `Test.@test` with `ArgCheck.@argcheck` to improve loading times by about 20 ms.

Here are some basic timings.

Master:
```julia-repl
julia> @time_imports import Accessors
      6.4 ms  MacroTools
               ┌ 0.0 ms InverseFunctions.__init__() 
     21.5 ms  InverseFunctions 89.30% compilation time
      0.1 ms  InverseFunctions → InverseFunctionsDatesExt
      0.2 ms  ConstructionBase
      0.1 ms  CompositionsBase
      0.1 ms  CompositionsBase → CompositionsBaseInverseFunctionsExt
     18.3 ms  Test
      0.1 ms  InverseFunctions → InverseFunctionsTestExt
               ┌ 0.0 ms Accessors.__init__() 
      8.6 ms  Accessors
```

This PR:
```julia-repl
julia> @time_imports import Accessors
      6.0 ms  MacroTools
               ┌ 0.0 ms InverseFunctions.__init__() 
     21.1 ms  InverseFunctions 89.71% compilation time
      0.1 ms  InverseFunctions → InverseFunctionsDatesExt
      0.3 ms  ConstructionBase
      0.1 ms  CompositionsBase
      0.1 ms  CompositionsBase → CompositionsBaseInverseFunctionsExt
      0.4 ms  ArgCheck
               ┌ 0.0 ms Accessors.__init__() 
      9.0 ms  Accessors
```

From what I can tell 
```julia
function test_getset_laws end
function test_modify_law end
function test_getsetall_laws end
function test_insertdelete_laws end
```

are all not documented, so changing them should not be a breaking change.